requirements.txt
streamlit>=1.28.0
openai>=1.3.0
PyPDF2>=3.0.1
python-dotenv>=1.0.0